specific int sizes

+++ _strings.h
@@ -13,15 +13,15 @@ typedef struct {
 
 
 _tokens* _string_tokenize(char* _str, char separator) {
-    int length = strlen(_str);
+    uint32_t length = strlen(_str);
 
-    int token_count = 0;
-    for(int i = 0; i < length; i++) {
+    uint32_t token_count = 0;
+    for(uint32_t i = 0; i < length; i++) {
         if (_str[i] == separator) {
             token_count++;
 
             // Ignore multiple separator characters in a row.
-            int j = i + 1;
+            uint32_t j = i + 1;
             while(1) {
                 if (_str[j] == separator) {
                     i++;
@@ -39,7 +39,7 @@ _tokens* _string_tokenize(char* _str, char separator) {
     toks->tokens = (char**)calloc(token_count, sizeof(char*));
 
     char* token;
-    for(int i = 0; i < token_count; i++) {
+    for(uint32_t i = 0; i < token_count; i++) {
         if (i == 0) {
             token = strtok(_str, &separator);
         } else {
@@ -61,7 +61,7 @@ int _tokens_comparator_descending(const void* tok1, const void* tok2) {
     return -1 * strcmp(*(const char**)tok1, *(const char**)tok2);
 }
 
-void _tokens_sort(_tokens* toks, int order) {
+void _tokens_sort(_tokens* toks, int8_t order) {
     if (order != 1 && order != -1) {
         errno = 22;
         perror("Invalid order value. order argument must be 1 (ascending) or -1 (descending)!");
@@ -76,7 +76,7 @@ void _tokens_sort(_tokens* toks, int order) {
 }
 
 void _tokens_free(_tokens* tokens) {
-    for(int i = 0; i < tokens->size; i++) {
+    for(uint32_t i = 0; i < tokens->size; i++) {
         free(tokens->tokens[i]);
     }
     free(tokens->tokens);
